Added pure_shell.go and renamed convertEnvToMap
